[441910]-Support for git-ref and sub directory is incomplete for installing Cordova plug-ins
CordovaPluginManager adds support for git-ref and subdirs that are
passed as part of the URI when installing plugins from git. With this
change the git variant installPlugin no longer needs the commit and
subdir explicitly as parameters. All the users of the API are also
adjusted to this change.

CordovaPluginSelectionPage also adds an info section that gives and
example of how to specify git-ref and subdir on a git url.