Bug 491580 - Implement Engines Management in config.xml editor
Fixes issues with tests running on platforms that do not have any cordova platforms installed.
Previously, the tests assumed that cordova-android 5.0.0 was installed; this has been changed
so that the tests should pass independent of the system's set up. Note that the changes
do result in the tests being more fragile, as they rely on side-effects of methods in
CordovaEngineProvider.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=491580
Signed-off-by: Angel Misevski <amisevsk@redhat.com>